Separate Python Scripts Into Libraries/Classes
CLStore was originally the only python script. After this commit, we now have cltools.py, clconfig.py, and clstore.py, each of which contain associated logic and code.

CLTools is designed to be the parent class with general functions. CLConfig is a class used to track and manage application configuration settings. And CLStore is the original application on its own (pulls and stores Craigslist listings for further analysis).