Add Support for Fetching Multiple Listings in CLData API
The Data() library in the CLData API has been modified to allow for fetching and returning multiple rows/listings. This will be helpful later on when performing searches or sorting a set of results.

Additionally, the CLData API has been changed to show the five msot recent listings when no listing ID is supplied as opposed to generating an error. This provides more robustness to the application.