Add Config Option for Log Verbosity
A config option has been added to CLStore in order to specify the minimum log verbosity threshold. Any log messages that are greater than the log level will be written; all others will be ignored.

This config option can be specified in the config file using the property name 'log_verbosity_lvl' or via a CLI option: -v or -verbosity.

Related To: Issue #16